refactor(memory): extract source test modules
- Move memory unit tests into sibling *_test.rs files
- Keep memory source modules focused on runtime behavior
